requirements-prod.txt
# Production dependencies - stable, pinned versions
pandas>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.26.0,<3.0.0
py-clob-client>=0.30.0,<1.0.0
tqdm>=4.60.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
apache-airflow>=2.8.0
websocket-client>=1.6.0